Commented out flex, bison, and m4 from each of the temp-system chapter files
These were added to the build because of the use of HJL binutils. If you
check the LFS earilier versions, you will notice these were only built during
the final system build. So we are going back to that methodology. 

Since the files are only commented out, it will be easy to re-enable them
if the reason should arise.

Flex, bison, and m4 are not needed for the final-system builds at all. They
are still built during the final-system stage.

Here is the proof that a lot people will be asking for
http://archives.linuxfromscratch.org/lfs-museum/5.1.1/LFS-BOOK-5.1.1-HTML/



git-svn-id: svn://svn.linuxfromscratch.org/LFS/branches/cross-lfs@5467 4aa44e1e-78dd-0310-a6d2-fbcd4c07a689