Info, script and data for svn migration.
The script is to create tracking branches for the corresponding svn
branches.  The author's transform file should be copied outside the
tree (or in the .git dir) and used for git fetches.  The path should
be put in the right place of the .git/config file (see the snippet
in this dir).